pyproject.toml
[tool.poetry]
name = "gui"
version = "0.1.0"
description = ""
authors = ["guilhermealbanaz <guibragatoalbnz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.1"
drf-spectacular = "^0.24.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
